Fix 32bit warnings

+++ src/mc-lib/lmap.h
@@ -25,13 +25,13 @@ typedef struct lm_s     lm_t;
 /**
 \brief A lattice is assumed to be represented by a 64bit pointer.
  */
-typedef uint64_t        lattice_t;
+typedef size_t          lattice_t;
 static const lattice_t  NULL_LATTICE = -1;
 
 /**
 \typedef The location of the iterator.
  */
-typedef uint64_t        lm_loc_t;
+typedef size_t          lm_loc_t;
 static const lm_loc_t   LM_NULL_LOC = -1;
 
 /**
